app/requirements.txt
langgraph
langchain
langchain-groq
langchain-community
python-dotenv
supabase
sentence_transformers
faiss-cpu
fastapi
uvicorn
sqlalchemy
pydantic
langgraph[sqlite]
pandas
numpy
python-docx
openpyxl
python-multipart
PyMuPDF  
pandas
python-docx